services/market_data_hub/requirements.txt
google-cloud-firestore
websockets
asyncio
